requirements.txt
matplotlib
numpy
pandas
Pillow>=7.1.0
plotly
seaborn
streamlit>=1.34.0
streamlit_option_menu
